requirements.txt
streamlit
python-dotenv
Pillow
PyMuPDF
google-generativeai